Bowtie parser and minor code tidy
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/ShortRead@34949 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ src/read_maq_map.cc
@@ -74,7 +74,7 @@ template< int max_readlen > SEXP read_maq_map_B( SEXP filename, SEXP maxreads )
         SET_STRING_ELT( seqnames, i, mkChar( mapheader->ref_name[i] ) );
     }
     if( INTEGER(maxreads)[0] < 0 || 
-        INTEGER(maxreads)[0] >= mapheader->n_mapped_reads )
+        INTEGER(maxreads)[0] >= (int) mapheader->n_mapped_reads )
         actnreads = mapheader->n_mapped_reads;
     else
         actnreads = INTEGER(maxreads)[0];
